refactor: extract map core CSS to styles folder
